Big Update 22/12/2019

+++ src/main/java/com/chika/server/controllers/RemoteIrController.java
@@ -1,27 +1,36 @@
 package com.chika.server.controllers;
 
 import com.chika.server.models.house.RemoteIr;
+import com.chika.server.payload.responses.ApiResponse;
+import com.chika.server.security.CurrentUser;
+import com.chika.server.security.UserPrincipal;
 import com.chika.server.services.IrValueService;
 import com.chika.server.services.RemoteIrService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
 /**
+ * To receive Remote Ir requests from client
  * @author Sy Nguyen
  * @version 1.0
- * @since 29-11-2019
+ * @since 22-12-2019
  */
 @RestController
-@RequestMapping("/remote-ir")
+@RequestMapping("/remoteir")
 public class RemoteIrController {
 
-    @Autowired
-    private RemoteIrService remoteIrService;
+    private final RemoteIrService remoteIrService;
 
-    @Autowired
-    private IrValueService irValueService;
+    private final IrValueService irValueService;
+
+    public RemoteIrController(RemoteIrService remoteIrService, IrValueService irValueService) {
+        this.remoteIrService = remoteIrService;
+        this.irValueService = irValueService;
+    }
 
     @GetMapping("/{roomId}")
     public List<RemoteIr> getAllRemotesByRoomId(@PathVariable String roomId) {
@@ -31,19 +40,27 @@ public class RemoteIrController {
     @PostMapping
     public RemoteIr saveRemote(@RequestBody RemoteIr remoteIr, @RequestParam("numOfButton") int numOfButton) {
         RemoteIr remote = remoteIrService.save(remoteIr);
-        irValueService.saveList(remote.getId(), numOfButton);
+        remote.setIrValues(irValueService.saveList(remote.getId(), numOfButton));
         return remote;
     }
 
     @PutMapping
-    public RemoteIr updateName(@RequestBody RemoteIr remoteIr) {
-        return remoteIrService.updateName(remoteIr.getId(), remoteIr.getName());
+    public ResponseEntity<?> updateName(@CurrentUser UserPrincipal currentUser, @RequestBody RemoteIr remoteIr) {
+        if (remoteIrService.isOwner(remoteIr.getId(), currentUser.getId())) {
+            return ResponseEntity.ok(remoteIrService.updateName(remoteIr.getId(), remoteIr.getName()));
+        }
+        return new ResponseEntity<>(new ApiResponse(false, "You are not remote owner"),
+                HttpStatus.BAD_REQUEST);
     }
 
     @DeleteMapping("/{id}")
-    public Boolean deleteRemote(@PathVariable String id) {
-        irValueService.deleteAllByRemoteIrId(id);
-        remoteIrService.deleteById(id);
-        return true;
+    public ResponseEntity<?> deleteRemote(@CurrentUser UserPrincipal currentUser, @PathVariable String id) {
+        if (remoteIrService.isOwner(id, currentUser.getId())) {
+            irValueService.deleteAllByRemoteIrId(id);
+            remoteIrService.deleteById(id);
+            return ResponseEntity.ok(new ApiResponse(true, "Remote has been deleted"));
+        }
+        return new ResponseEntity<>(new ApiResponse(false, "You are not remote owner"),
+                HttpStatus.BAD_REQUEST);
     }
 }
